LayoutTests/http/tests/resourceLoadStatistics/partitioned-and-unpartitioned-cookie-with-partitioning-timeout-expected.txt

CONSOLE MESSAGE: line 117: Bootstrapping test.
CONSOLE MESSAGE: line 68: Step 1.
CONSOLE MESSAGE: line 73: Step 2.
CONSOLE MESSAGE: line 80: Step 3.
CONSOLE MESSAGE: line 87: Step 4.
CONSOLE MESSAGE: line 57: setNormalCookiePartitioningTimeOutPlusUserInteractionAndContinue().
CONSOLE MESSAGE: line 94: Step 5.
CONSOLE MESSAGE: line 51: setShortCookiePartitioningTimeOutPlusFireShouldPartitionCookiesHandlerAndContinue().
CONSOLE MESSAGE: line 46: fireShouldPartitionCookiesHandlerAndContinue().
CONSOLE MESSAGE: line 101: Step 6.
CONSOLE MESSAGE: line 57: setNormalCookiePartitioningTimeOutPlusUserInteractionAndContinue().
CONSOLE MESSAGE: line 108: Step 7.
PASS successfullyParsed is true

TEST COMPLETE
  

--------
Frame: '<!--frame1-->'
--------
Should receive no cookies.
Did not receive cookie named 'firstPartyCookie'.
Did not receive cookie named 'thirdPartyCookie'.
Client-side document.cookie:

--------
Frame: '<!--frame2-->'
--------
Setting partitioned, third-party cookie.


--------
Frame: '<!--frame3-->'
--------
Should only receive partitioned, third-party cookie.
Did not receive cookie named 'firstPartyCookie'.
Received cookie named 'thirdPartyCookie'.
Client-side document.cookie: thirdPartyCookie=value

--------
Frame: '<!--frame4-->'
--------
After first user interaction, should still only receive a partitioned, third-party cookie.
Did not receive cookie named 'firstPartyCookie'.
Received cookie named 'thirdPartyCookie'.
Client-side document.cookie: thirdPartyCookie=value

--------
Frame: '<!--frame5-->'
--------
Should only receive partitioned, third-party cookie.
Did not receive cookie named 'firstPartyCookie'.
Received cookie named 'thirdPartyCookie'.
Client-side document.cookie: thirdPartyCookie=value

--------
Frame: '<!--frame6-->'
--------
After second user interaction, should still only receive a partitioned, third-party cookie.
Did not receive cookie named 'firstPartyCookie'.
Received cookie named 'thirdPartyCookie'.
Client-side document.cookie: thirdPartyCookie=value
